chore: move audit structure back to team 2
